% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_performance.R
\name{calc_bfdr}
\alias{calc_bfdr}
\title{Calculate the estimated false discovery rate}
\usage{
calc_bfdr(p)
}
\arguments{
\item{p}{A vector of posterior probabilities}
}
\description{
Calculate the estimated false discovery rate
}
