% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.helpers-package.R
\docType{package}
\name{de.helpers-package}
\alias{de.helpers}
\alias{de.helpers-package}
\title{de.helpers: Functions to help with my differential expression and differential translational efficiency projects}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dcannonwalker/de.helpers}
  \item Report bugs at \url{https://github.com/dcannonwalker/de.helpers/issues}
}

}
\author{
\strong{Maintainer}: David Walker \email{dcwalker@iastate.edu}

}
\keyword{internal}
