% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_performance.R
\name{.fawcett1_internal}
\alias{.fawcett1_internal}
\title{Do the inside step of fawcett1}
\usage{
.fawcett1_internal(R, tnsort, i = 1, FP = 0, TP = 0, N, P)
}
\description{
Do the inside step of fawcett1
}
