% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_performance.R
\name{.fawcett2_internal}
\alias{.fawcett2_internal}
\title{Do the inside of fawcett2}
\usage{
.fawcett2_internal(A = 0, tnsort, i = 1, FP = 0, TP = 0, N, P)
}
\description{
Do the inside of fawcett2
}
