% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_parameters.R
\name{estimate_parameters}
\alias{estimate_parameters}
\title{Estimate parameters from a real rna-seq data set}
\usage{
estimate_parameters(counts, group, min_count = 10, filter_fn = sum)
}
\arguments{
\item{counts}{a matrix of counts}

\item{group}{a vector of treatment group labels corresponding to
columns of \code{counts}}

\item{min_count}{the minimum count to keep a tag in the dataset}

\item{filter_fn}{function to apply rowwise; result is compared to \code{min_count}}
}
\description{
Estimate parameters from a real rna-seq data set
}
