% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_performance.R
\name{fawcett1}
\alias{fawcett1}
\title{Algorithm 1 from Fawcett 2006, for generating ROC points}
\usage{
fawcett1(p, tn)
}
\arguments{
\item{p}{Vector of null probabilities}

\item{tn}{Vector of actual classes}
}
\description{
Algorithm 1 from Fawcett 2006, for generating ROC points
}
