% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_performance.R
\name{fawcett3}
\alias{fawcett3}
\title{Algorithm 3 from Fawcett 2006; vertical averaging of ROC curves}
\usage{
fawcett3(x0, rocs, average = TRUE)
}
\arguments{
\item{x0}{A vector of fpr values}

\item{rocs}{A list of ROC curves}
}
\description{
Algorithm 3 from Fawcett 2006; vertical averaging of ROC curves
}
