% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative_methods.R
\name{fit_edgeR}
\alias{fit_edgeR}
\title{Fit a standard \code{edgeR} pipeline for two-group comparison
Note to self: to be extended to more general designs}
\usage{
fit_edgeR(counts, design, ...)
}
\arguments{
\item{counts}{A matrix of counts; columns are samples, rows are tags}

\item{design}{A design matrix for the columns of \code{counts}}

\item{...}{Additional parameters to pass on}
}
\description{
Fit a standard \code{edgeR} pipeline for two-group comparison
Note to self: to be extended to more general designs
}
