% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative_methods.R
\name{fit_limma}
\alias{fit_limma}
\title{Fit a standard \code{limma-trend} or \code{limma-voom} pipeline for
two-group comparison
Note to self: to be extended to more general designs
Note that fdr is 'BH' adjustment to p-value}
\usage{
fit_limma(counts, design, use_voom = TRUE, ...)
}
\arguments{
\item{counts}{A matrix of counts; columns are samples, rows are tags}

\item{design}{A design matrix for the columns of \code{counts}}

\item{use_voom}{Logical - use \code{limma-voom}? Otherwise use \code{limma-trend}}

\item{...}{Additional parameters to pass on}
}
\description{
Fit a standard \code{limma-trend} or \code{limma-voom} pipeline for
two-group comparison
Note to self: to be extended to more general designs
Note that fdr is 'BH' adjustment to p-value
}
