% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative_methods.R
\name{fit_ngstan}
\alias{fit_ngstan}
\title{Fit a standard \code{ngstan} pipeline for two group comparison}
\usage{
fit_ngstan(counts, design, ...)
}
\arguments{
\item{counts}{A matrix of counts; columns are samples, rows are tags}

\item{design}{A design matrix for the columns of \code{counts}}

\item{...}{Additional parameters to pass on}
}
\description{
Fit a standard \code{ngstan} pipeline for two group comparison
}
