% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_filesystem.R
\name{generate_id}
\alias{generate_id}
\title{Generate an identifier}
\usage{
generate_id(n_ids = 1, n_char = 10, ids_to_check = NULL, seed = NULL, i = 0)
}
\arguments{
\item{n_ids}{Number of ids to generate}

\item{n_char}{Number of characters}

\item{ids_to_check}{A vector of other ids to check against}

\item{seed}{Seed value to set}
}
\description{
Generate an identifier
}
