% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngstan.R
\name{get_contrast_posterior_mean}
\alias{get_contrast_posterior_mean}
\title{Functions to wrangle \code{{ngstan}} that I'm not ready to put into
the package itself yet}
\usage{
get_contrast_posterior_mean(contrast, comps, d_pmf, beta)
}
\arguments{
\item{contrast}{A vector giving the contrast}

\item{comps}{The \code{comps} matrix taken from \code{standata}}

\item{d_pmf}{The \code{d_pmf} draws}

\item{beta}{The \code{beta} draws}
}
\description{
Note that \code{posterior::extract_variable_array()} returns an array that
is draws x chains x tags (where applicable) x variable
}
\details{
Get the posterior mean for a contrast
}
