% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_filesystem.R
\name{get_curve_path}
\alias{get_curve_path}
\title{Get the standard file path for saved ROC or FDR curve data}
\usage{
get_curve_path(simulation_id, root, method, type)
}
\arguments{
\item{simulation_id}{The simulation study identifier}

\item{root}{The parent directory for all simulation studies}

\item{method}{The name of the method whose fits we want to retrieve}

\item{type}{ROC, FDR, or both?}
}
\description{
Get the standard file path for saved ROC or FDR curve data
}
