% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchers.R
\name{launch_simulation_study}
\alias{launch_simulation_study}
\title{Launch model fits for a set of methods for all data sets
in a given simulation study}
\usage{
launch_simulation_study(
  simulation_id,
  methods = c("ngstan", "edgeR", "DESeq2", "limma"),
  ...,
  verbose = TRUE,
  root = file.path("out", "simulation_studies"),
  script_name = "slurm/single_method_launcher",
  command = "sbatch",
  slurm = TRUE
)
}
\description{
Launch model fits for a set of methods for all data sets
in a given simulation study
}
