% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_study_analysis.R
\name{make_average_fdr_curve}
\alias{make_average_fdr_curve}
\title{Construct the average FDR curve for a set of model fits}
\usage{
make_average_fdr_curve(method_data, x0)
}
\arguments{
\item{method_data}{List of model fits for a single method}

\item{x0}{A vector of fpr or nominal fdr values at which to take
the averages of the ROC or FDR curves}
}
\description{
Construct the average FDR curve for a set of model fits
}
