% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_filesystem.R
\name{make_dataset_dirs}
\alias{make_dataset_dirs}
\title{Make data set directories for all the data set ids in
\code{dataset_ids.txt} for a given simulation study}
\usage{
make_dataset_dirs(simulation_id, root = file.path("out", "simulation_studies"))
}
\arguments{
\item{simulation_id}{The simulation study identifier}

\item{root}{The parent directory for all simulation studies}
}
\description{
Make data set directories for all the data set ids in
\code{dataset_ids.txt} for a given simulation study
}
