% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_study_analysis.R
\name{make_fdr_curve}
\alias{make_fdr_curve}
\title{Construct an FDR curve for a given model fit}
\usage{
make_fdr_curve(method_data, misleading_names = FALSE)
}
\arguments{
\item{method_data}{Model fit for a single method,
as read in by \code{read_method_data()}}

\item{misleading_names}{Output will have misleading column names
so that it can be passed to \code{fawcett3()} without having to change
the column names later (or that function to be more generic);
change it later if you feel like it}
}
\description{
Construct an FDR curve for a given model fit
}
