% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_study_analysis.R
\name{make_roc_curve}
\alias{make_roc_curve}
\title{Construct an ROC curve for a given model fit}
\usage{
make_roc_curve(method_data)
}
\arguments{
\item{method_data}{Model fit for a single method,
as read in by \code{read_method_data()}}
}
\description{
Construct an ROC curve for a given model fit
}
