% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_filesystem.R
\name{make_sim_dir}
\alias{make_sim_dir}
\title{Set up a directory to contain all files for a given simulation study}
\usage{
make_sim_dir(simulation_id, root = file.path("out", "simulation_studies"))
}
\arguments{
\item{simulation_id}{The simulation study identifier}

\item{root}{The parent directory for all simulation studies}
}
\description{
Set up a directory to contain all files for a given simulation study
}
