% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_filesystem.R
\name{plot_curves}
\alias{plot_curves}
\title{Create plots of ROC and FDR curves}
\usage{
plot_curves(
  simulation_id,
  root = "out/simulation_studies",
  method = c("edgeR", "DESeq2", "limma", "ngstan"),
  type = c("roc", "fdr")
)
}
\arguments{
\item{simulation_id}{The simulation study identifier}

\item{root}{The parent directory for all simulation studies}

\item{method}{The name of the method whose fits we want to retrieve}
}
\description{
Create plots of ROC and FDR curves
}
