% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchers.R
\name{prepare_slurm_prompt}
\alias{prepare_slurm_prompt}
\title{Prepare a system prompt}
\usage{
prepare_slurm_prompt(
  simulation_id,
  dataset_id,
  method,
  slurm_script_name,
  ...,
  root = "out/simulation_studies"
)
}
\description{
Prepare a system prompt
}
