% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_filesystem.R
\name{read_method_data}
\alias{read_method_data}
\title{Read in all the saved fits for a given simulation study and method}
\usage{
read_method_data(
  simulation_id,
  root = "out/simulation_studies",
  method = c("edgeR", "DESeq2", "limma", "ngstan")
)
}
\arguments{
\item{simulation_id}{The simulation study identifier}

\item{root}{The parent directory for all simulation studies}

\item{method}{The name of the method whose fits we want to retrieve}
}
\description{
Read in all the saved fits for a given simulation study and method
}
