% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_filesystem.R
\name{remove_datasets}
\alias{remove_datasets}
\title{Remove all existing data set folders from a simulation study}
\usage{
remove_datasets(simulation_id, root = file.path("out", "simulation_studies"))
}
\arguments{
\item{simulation_id}{The simulation study identifier}

\item{root}{The parent directory for all simulation studies}
}
\description{
Remove all existing data set folders from a simulation study
}
