% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_method.R
\name{run_method}
\alias{run_method}
\title{Fit a model to saved simulated data}
\usage{
run_method(
  simulation_id,
  dataset_id,
  method,
  root = file.path("out", "simulation_studies")
)
}
\arguments{
\item{simulation_id}{The simulation study identifier}

\item{dataset_id}{The data set identifier}

\item{root}{The parent directory for all simulation studies}
}
\description{
Fit a model to saved simulated data
}
