% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_counts.R
\name{sample_dispersion}
\alias{sample_dispersion}
\title{General outline:
A generic for simulating counts, paired with S4 classes for
each class of input parameters
Sample a dispersion given a mean and a set of mean and dispersion pairs}
\usage{
sample_dispersion(x, mean_pars, dispersion_pars, interval = 20)
}
\arguments{
\item{x}{Given mean}

\item{mean_pars}{Vector of estimated means}

\item{dispersion_pars}{Vector of estimated dispersions}

\item{interval}{Only pairs where \code{abs(mean_pars - x) < interval} will
be used}
}
\description{
General outline:
A generic for simulating counts, paired with S4 classes for
each class of input parameters
Sample a dispersion given a mean and a set of mean and dispersion pairs
}
