% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_counts.R
\name{sample_neighborhood}
\alias{sample_neighborhood}
\title{General outline:
A generic for simulating counts, paired with S4 classes for
each class of input parameters
Sample from a neighborhood around a given value}
\usage{
sample_neighborhood(x0, x, y, interval = 3)
}
\arguments{
\item{x0}{Given value to compare to \code{x}}

\item{x}{Vector from which to get neighborhood indices}

\item{y}{Vector to sample from}

\item{interval}{Only pairs where \code{abs(x - x0) < interval} will
be used}
}
\description{
General outline:
A generic for simulating counts, paired with S4 classes for
each class of input parameters
Sample from a neighborhood around a given value
}
