% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_filesystem.R
\name{save_curve_plots}
\alias{save_curve_plots}
\title{Create and save ROC and FDR curves for a simulation study}
\usage{
save_curve_plots(
  simulation_id,
  root = "out/simulation_studies",
  method = c("edgeR", "DESeq2", "limma", "ngstan"),
  type = c("roc", "fdr")
)
}
\arguments{
\item{simulation_id}{The simulation study identifier}

\item{root}{The parent directory for all simulation studies}

\item{method}{The name of the method whose fits we want to retrieve}
}
\description{
Create and save ROC and FDR curves for a simulation study
}
