% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_filesystem.R
\name{save_datasets}
\alias{save_datasets}
\title{Save simulated data sets for a simulation study}
\usage{
save_datasets(
  simulation_id,
  sim_data,
  design,
  root = file.path("out", "simulation_studies")
)
}
\arguments{
\item{simulation_id}{The simulation study identifier}

\item{sim_data}{A list of simulated data sets, each element named
matching one of the data set ids in the simulation study}

\item{design}{The design matrix for the simulated counts}

\item{root}{The parent directory for all simulation studies}
}
\description{
Save simulated data sets for a simulation study
}
