% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_filesystem.R
\name{save_method_curves}
\alias{save_method_curves}
\title{A wrapper to create and save average ROC and FDR curves}
\usage{
save_method_curves(
  simulation_id,
  root = "out/simulation_studies",
  method = c("edgeR", "DESeq2", "limma", "ngstan"),
  type = c("roc", "fdr"),
  x0 = seq(0, 1, length = 100)
)
}
\arguments{
\item{simulation_id}{The simulation study identifier}

\item{root}{The parent directory for all simulation studies}

\item{method}{The name of the method whose fits we want to retrieve}

\item{type}{ROC, FDR, or both?}

\item{x0}{A vector of fpr or nominal fdr values at which to take
the averages of the ROC or FDR curves}
}
\description{
A wrapper to create and save average ROC and FDR curves
}
\section{Running and analyzing simulation studies}{

To use the helpers to run and analyze a simulation study, you're intended
to follow these steps in order:
\enumerate{
\item Simulate data sets and save them
\item Fit the set of methods and save their outputs for each data set
\item Create tables for average ROC and FDR curves and save them
\item Plot the ROC and FDR curves and save the plots
}
}

\section{Simulate data}{

\code{generate_id()}, \code{make_sim_dir()}, \code{make_dataset_dir()}, \code{simulate_counts_from_dataset()},
\code{save_datasets()}
}

\section{Fit methods}{

\code{run_method()}
}

\section{Create summary tables}{

\code{save_method_curves()}
}

\section{Save plots}{

\code{save_curve_plots()}
}

