% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_counts.R
\name{simulate_counts}
\alias{simulate_counts}
\title{Simulate RNASeq counts}
\usage{
simulate_counts(
  mean_pars,
  dispersion_pars,
  offset_pars,
  n_tags,
  design,
  offset_options = list(),
  effects_options = list(),
  dispersion_interval = log(20)
)
}
\arguments{
\item{mean_pars}{A matrix of per-tag regression coefficients,
with the first column the intercept}

\item{dispersion_pars}{A vector of estimated per-tag dispersions}

\item{offset_pars}{A vector of estimated sample offsets}

\item{n_tags}{The number of tags to simulate}

\item{design}{The design matrix for the simulated samples - should have
two columns, with the first column the intercept}

\item{offset_options}{A list of options to control the behavior of
\code{simulate_offsets()}}

\item{effects_options}{A list of options to control the behavior of
\code{simulate_effects()}}

\item{dispersion_interval}{After simulating the base mean \code{b0},
the dispersion is sampled from
\code{dispersion_pars[abs(mean_pars[, 1] - b0) < dispersion_interval]}}
}
\description{
Simulate RNASeq counts
}
