% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_counts.R
\name{simulate_effects.dst}
\alias{simulate_effects.dst}
\title{Not really supported yet...}
\usage{
simulate_effects.dst(
  n_tags,
  theta,
  de_prob,
  distribution = c("exponential"),
  ...
)
}
\arguments{
\item{theta}{Exponential parameter to be used if \code{method} is \code{"dst"}}

\item{de_prob}{Either a single probability representing the chance
that any given tag exhibits differential expression,
or a named vector of two probabilities, representing the chance
that any given tag exhibits differential expression in the \code{up} or \code{down}
direction; if a single probability, half of differentially expressed tags
will be \code{up} regulated; used only if \code{method} is \code{"dst"}}

\item{min_log_fc}{Minimum value added to draws from \code{"dst"}}
}
\description{
Not really supported yet...
}
