% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_counts.R
\name{simulate_effects.emp}
\alias{simulate_effects.emp}
\title{In development...
Simulate effects for a simple two-group design
by sampling from estimated coefficients}
\usage{
simulate_effects.emp(
  n_tags,
  mean_pars,
  p_null,
  b1_min = 0,
  keep_tag_together = FALSE,
  ...
)
}
\arguments{
\item{n_tags}{The number of tags to simulate}

\item{...}{Arguments passed to sub-functions}
}
\description{
In development...
Simulate effects for a simple two-group design
by sampling from estimated coefficients
}
