% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_counts.R
\name{simulate_effects}
\alias{simulate_effects}
\title{Simulate treatment effects}
\usage{
simulate_effects(n_tags, ...)
}
\arguments{
\item{n_tags}{The number of tags to simulate}

\item{...}{Arguments passed to sub-functions}

\item{method}{One of \code{c("emp", "dst")}, for sampling from estimated
coefficients or sampling from an exponential; if \code{method} is \code{"emp"},
\code{mean_pars} must not be \code{NULL}}
}
\description{
Simulate treatment effects
}
