% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_counts.R
\name{simulate_offsets}
\alias{simulate_offsets}
\title{Simulate per-sample offsets}
\usage{
simulate_offsets(offset_pars, n_samples, method = c("default"))
}
\arguments{
\item{offset_pars}{A vector of estimated offset parameters}

\item{n_samples}{The number of samples for which to simulate offsets}
}
\description{
Simulate per-sample offsets
}
