% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_performance.R
\name{tpr_for_fpr}
\alias{tpr_for_fpr}
\title{For a given fpr, find the fpr values on the ROC curve that bracket it}
\usage{
tpr_for_fpr(roc, x0)
}
\description{
For a given fpr, find the fpr values on the ROC curve that bracket it
}
