% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_performance.R
\name{trapezoid_area}
\alias{trapezoid_area}
\title{Calculate the area of a trapezoidal slice
of the AUC}
\usage{
trapezoid_area(x1, x2, y1, y2)
}
\description{
Calculate the area of a trapezoidal slice
of the AUC
}
