% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alternative_methods.R
\name{fit_ngstan}
\alias{fit_ngstan}
\title{Fit a standard \code{ngstan} pipeline for two group comparison}
\usage{
fit_ngstan(
  counts,
  design,
  mixture_probability = 0.8,
  iter_warmup = 1000,
  iter_sampling = 1000,
  parallel_chains = 4,
  grainsize = NULL,
  use_multithread = FALSE,
  save_fit_raw = TRUE,
  ...
)
}
\arguments{
\item{counts}{A matrix of counts; columns are samples, rows are tags}

\item{design}{A design matrix for the columns of \code{counts}}

\item{mixture_probability}{The prior probability of null differential expression}

\item{iter_warmup}{Warmup iterations for Stan model}

\item{iter_sampling}{Sampling iterations for Stan model}

\item{parallel_chains}{Number of parallel chains for Stan model}

\item{grainsize}{Grainsize for multi-thread Stan model}

\item{use_multithread}{Use the multithread model?}

\item{save_fit_raw}{Save the full cmdstan output?}

\item{...}{Additional parameters to pass on}
}
\description{
Fit a standard \code{ngstan} pipeline for two group comparison
}
