% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchers.R
\name{launch_single_method}
\alias{launch_single_method}
\title{Launch single method using system prompt}
\usage{
launch_single_method(
  simulation_id,
  dataset_id,
  method,
  ...,
  root = "out/simulation_studies"
)
}
\description{
For \code{ngstan}, the first two elements of \code{...} will be used
as \code{iter_warmup} and \code{iter_sampling}
}
