% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_filesystem.R
\name{make_dataset_dir}
\alias{make_dataset_dir}
\title{Set up a directory to contain all the files for a single simulated data set
within a simulation study}
\usage{
make_dataset_dir(
  dataset_id,
  simulation_id,
  root = file.path("out", "simulation_studies"),
  doc = TRUE
)
}
\arguments{
\item{dataset_id}{The data set identifier}

\item{simulation_id}{The simulation study identifier}

\item{root}{The parent directory for all simulation studies}

\item{doc}{Add the dataset id to \code{dataset_ids.txt}?}
}
\description{
Set up a directory to contain all the files for a single simulated data set
within a simulation study
}
