% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_filesystem.R
\name{save_method_curves}
\alias{save_method_curves}
\title{A wrapper to create and save average ROC and FDR curves}
\usage{
save_method_curves(
  simulation_id,
  root = "out/simulation_studies",
  method = c("edgeR", "DESeq2", "limma", "ngstan"),
  x0 = seq(0, 1, length = 100)
)
}
\arguments{
\item{simulation_id}{The simulation study identifier}

\item{root}{The parent directory for all simulation studies}

\item{method}{The name of the method whose fits we want to retrieve}

\item{x0}{A vector of fpr or nominal fdr values at which to take
the averages of the ROC or FDR curves}
}
\description{
A wrapper to create and save average ROC and FDR curves
}
