% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_counts.R
\name{simulate_effects.emp}
\alias{simulate_effects.emp}
\title{Simulate effects by sampling from estimated coefficients}
\usage{
simulate_effects.emp(n_tags, mean_pars, interval = 1)
}
\arguments{
\item{n_tags}{The number of tags to simulate}

\item{mean_pars}{A matrix of per-tag regression coefficients - should have
two columns, with the first column the intercept}
}
\description{
Simulate effects by sampling from estimated coefficients
}
