% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_counts.R
\name{simulate_effects}
\alias{simulate_effects}
\title{Simulate treatment effects}
\usage{
simulate_effects(n_tags, ...)
}
\arguments{
\item{n_tags}{The number of tags to simulate}

\item{...}{Arguments passed to sub-functions}

\item{method}{One of \code{c("emp", "emp.paired")}, for two group or
two group with paired samples}
}
\description{
Simulate treatment effects
}
