% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_counts.R
\name{simulate_offsets}
\alias{simulate_offsets}
\title{Simulate per-sample offsets}
\usage{
simulate_offsets(offset_pars, n_samples, n_tags, method = c("default"))
}
\arguments{
\item{offset_pars}{A vector of estimated sample offsets}

\item{n_samples}{The number of samples for which to simulate offsets}

\item{n_tags}{The number of tags to simulate}

\item{method}{One of \code{c("emp", "emp.paired")}, for two group or
two group with paired samples}
}
\description{
Simulate per-sample offsets
}
